third_party/WebKit/LayoutTests/http/tests/html/focus-pending-stylesheets.html

<!doctype html>
<head>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
  <script>
    'use strict';

    document.addEventListener("DOMContentLoaded", function() {
      test(() => {
        var editableText = document.getElementById("in");
        editableText.focus();
        assert_equals(document.activeElement, editableText);
      }, '<div> should be focusable even when there are pending stylesheets');
    });
  </script>
  <link rel="stylesheet" type="text/css" href="/resources/load-and-stall.php?name=dummy.css&stallAt&stallFor=2">
</head>
<body>
  <div contenteditable id="in"></div>
</body>
